requirements.txt
absl-py==1.4.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.5.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bddl==1.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==1.6.0
cmake==3.27.2
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.6
decorator==4.4.2
deepspeed==0.11.1
defusedxml==0.7.1
diffusers==0.31.0
Django==4.2.16
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.10
egl-probe==1.0.2
einops==0.6.1
einops-exts==0.0.4
ema-pytorch==0.2.3
environs==9.5.0
etils==1.3.0
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
flow-vis==0.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
future==0.18.2
gast==0.4.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.43
glfw==2.6.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
gym==0.26.0
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.9.0
hjson==3.1.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
image==1.5.33
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.27.1
ipython==8.12.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.14
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupytext==1.16.4
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
-e git+https://github.com/Lifelong-Robot-Learning/LIBERO.git@f78abd68ee283de9f9be3c8f7e2a9ad60246e95c#egg=libero
lightning==2.1.0
lightning-utilities==0.9.0
line-profiler==4.1.1
lit==16.0.6
llvmlite==0.40.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
mediapy==1.2.0
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.7
multidict==6.1.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
proglog==0.1.10
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybullet-svl==3.1.6.4
pycparser==2.22
pydantic==1.10.18
Pygments==2.18.0
pynput==1.7.6
pynvml==11.5.3
PyOpenGL==3.1.7
pyparsing==3.0.9
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-xlib==0.33
pytorch-lightning==2.1.0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
# robomimic @ file:///mnt/code/ZhangXiaoxiong/ATM/third_party/robomimic
# -e git+https://github.com/ARISE-Initiative/robosuite.git@eafb81f54ffc104f905ee48a16bb15f059176ad3#egg=robosuite&subdirectory=../../../third_party/robosuite
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.3
sentry-sdk==2.16.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
terminado==0.18.1
tf-keras==2.15.0
tfds-nightly==4.9.2.dev202308090034
threadpoolctl==3.5.0
timm==0.9.6
tinycss2==1.3.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.2
torch==2.0.1
torchmetrics==1.4.2
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.31.0
triton==2.0.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.5.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wandb==0.15.9
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.14.0
zipp==3.20.2
